Added DC blocker (high pass filter) to DiscreteTrigger.

+++ Source/DrumKit/Triggers/TriggerState.h
@@ -18,7 +18,7 @@ namespace DrumKit
 		int sensorId;
 		float value;
 		bool isTrig;
-		long long trigTime;
+		int64_t trigTime;
 
 	};
